Module path mappings for Jest. Reinstate AppProviders for test entrypoint
